refactor: Renamed script_test_helper to parser_test_helper
