+++ test/recipe_parser_tests/kptncook_test.dart
@@ -7,7 +7,7 @@ import 'package:ingredient_collector/src/recipe_controller/recipe_controller.dar
 import 'package:ingredient_collector/src/recipe_parser/recipe_parser.dart'
     show KptnCookParser;
 
-import 'script_test_helper.dart';
+import 'parser_test_helper.dart';
 
 void main() {
   setUp(() => RecipeCache().cache.clear());
